requirements.txt
asgiref==3.8.1
attrs==25.1.0
concurrency==0.1.0
coverage==7.6.10
Django==4.2.18
django-auto-prefetching==0.2.12
django-concurrency==2.6.0
django-filter==24.3
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
py==1.11.0
PyJWT==2.10.1
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-profiling==1.7.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
rules==3.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
